Fix models

+++ wallet/models.py
@@ -3,4 +3,4 @@ from django.db import models
 
 class Wallet(models.Model):
     label = models.CharField(max_length=200)
-    balance = models.PositiveIntegerField()
+    balance = models.DecimalField(max_digits=65, decimal_places=18)
